refactor(art-service): consolidate admin art repository into IArtPostRepository
Remove IAdminArtRepository and merge its functionality into IArtPostRepository to reduce code duplication and simplify the repository layer. Update all dependent use cases and DI configuration to use the consolidated repository.
