+++ services/art-service/src/application/usecase/admin/GetArtStatsUseCase.ts
@@ -1,12 +1,12 @@
 import { injectable, inject } from "inversify";
 import { TYPES } from "../../../infrastructure/Inversify/types";
-import { IAdminArtRepository } from "../../../domain/repositories/IAdminArtRepository";
+import { IArtPostRepository } from "../../../domain/repositories/IArtPostRepository";
 import { IGetArtStatsUseCase } from "../../interface/usecase/admin/IGetArtStatsUseCase";
 
 @injectable()
 export class GetArtStatsUseCase implements IGetArtStatsUseCase {
   constructor(
-    @inject(TYPES.IAdminArtRepository) private _repository: IAdminArtRepository
+    @inject(TYPES.IArtPostRepository) private _repository: IArtPostRepository
   ) {}
 
   async execute(): Promise<{
