+++ services/art-service/src/application/usecase/admin/GetCategoryStatsUseCase.ts
@@ -1,12 +1,12 @@
 import { injectable, inject } from "inversify";
 import { TYPES } from "../../../infrastructure/Inversify/types";
-import { IAdminArtRepository } from "../../../domain/repositories/IAdminArtRepository";
+import { IArtPostRepository } from "../../../domain/repositories/IArtPostRepository";
 import { IGetCategoryStatsUseCase } from "../../interface/usecase/admin/IGetCategoryStatsUseCase";
 
 @injectable()
 export class GetCategoryStatsUseCase implements IGetCategoryStatsUseCase {
   constructor(
-    @inject(TYPES.IAdminArtRepository) private _repository: IAdminArtRepository
+    @inject(TYPES.IArtPostRepository) private _repository: IArtPostRepository
   ) {}
 
   async execute(): Promise<{ category: string; count: number }[]> {
