+++ services/art-service/src/application/usecase/admin/GetTopArtsUseCase.ts
@@ -1,13 +1,13 @@
 import { injectable, inject } from "inversify";
 import { TYPES } from "../../../infrastructure/Inversify/types";
-import { IAdminArtRepository } from "../../../domain/repositories/IAdminArtRepository";
+import { IArtPostRepository } from "../../../domain/repositories/IArtPostRepository";
 import { IGetTopArtsUseCase } from "../../interface/usecase/admin/IGetTopArtsUseCase";
 import { ArtPost } from "../../../domain/entities/ArtPost";
 
 @injectable()
 export class GetTopArtsUseCase implements IGetTopArtsUseCase {
   constructor(
-    @inject(TYPES.IAdminArtRepository) private _repository: IAdminArtRepository
+    @inject(TYPES.IArtPostRepository) private _repository: IArtPostRepository
   ) {}
 
   async execute(limit: number, type: 'likes' | 'price'): Promise<ArtPost[]> {
