+++ services/art-service/src/domain/repositories/IArtPostRepository.ts
@@ -1,7 +1,17 @@
-import { ArtPost } from "../../domain/entities/ArtPost";
+import { ArtPost, PostStatus, PostType, PriceType } from "../../domain/entities/ArtPost";
 import { IBaseRepository } from "./IBaseRepository";
 
+export interface AdminArtFilters {
+  status?: PostStatus;
+  postType?: PostType;
+  priceType?: PriceType;
+  search?: string;
+  userId?: string;
+  artIds?: string[]; // For Elasticsearch integration
+}
+
 export interface IArtPostRepository extends IBaseRepository<ArtPost> {
+  // User-facing methods
   getAllArt(page: number, limit: number): Promise<any>;
   findById(postId: string): Promise<any>;
   findByArtName(artName: string): Promise<any>;
@@ -27,4 +37,22 @@ export interface IArtPostRepository extends IBaseRepository<ArtPost> {
     page: number,
     limit: number
   ): Promise<ArtPost[]>;
+
+  // Admin methods (merged from IAdminArtRepository)
+  findAll(
+    page: number,
+    limit: number,
+    filters?: AdminArtFilters
+  ): Promise<{ arts: any[]; total: number }>;
+  
+  countStats(): Promise<{
+    total: number;
+    free: number;
+    premium: number;
+    aiGenerated: number;
+  }>;
+  
+  updateStatus(id: string, status: PostStatus): Promise<ArtPost | null>;
+  getTopArts(limit: number, type: 'likes' | 'price'): Promise<ArtPost[]>;
+  getCategoryStats(): Promise<{ category: string; count: number }[]>;
 }
